requirements.txt
Django==1.10.6
requests==2.13.0
psycopg2
django-filter==1.0.2
djangorestframework==3.6.2
django-model-changes
celery[sqlalchemy,gevent,redis]
Markdown==2.6.8
coreapi==2.3.0
drfdocs==0.0.11
redis==2.10.5
mock
django-celery-results
dnspython
